mypy
src/spei/plot.py:58: error: Argument 2 to "plot" of "Axes" has incompatible type "ExtensionArray | ndarray[Any, Any]"; expected "float | _SupportsArray[dtype[Any]] | _NestedSequence[_SupportsArray[dtype[Any]]] | bool | int | float | complex | str | bytes | _NestedSequence[bool | int | float | complex | str | bytes] | str"
 [arg-type]
Found 1 error in 1 file (checked 8 source files)

+++ src/spei/plot.py
@@ -55,7 +55,7 @@ def si(
         colormap = plt.get_cmap(cmap)
 
     if background:
-        ax.plot(si.index, si.values, linewidth=0.8, color="k")
+        ax.plot(si.index, si.values.astype(float), linewidth=0.8, color="k")
         ax.axhline(0, linestyle="--", linewidth=0.5, color="k")
 
         droughts = si.to_numpy(dtype=float, copy=True)
